--- JaeHyeonLee/BOJ/boj10870_B2_피보나치수5.py
@@ -1,13 +1,0 @@
-import sys
-
-input = sys.stdin.readline
-
-n = int(input())
-
-
-def fibo(num):
-    if num == 0 or num == 1:
-        return num
-    return fibo(num - 1) + fibo(num - 2)
-
-print(fibo(n))
