+++ JaeHyeonLee/README.md
@@ -1,4 +1,3 @@
 # 1114
 1. 함수
   - 로직을 담아서 반복 사용
-2.
